@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import './models/category.dart';
 import './models/meal.dart';
 
-const DUMMY_CATEGORIES = const [
+const dummyCategories = [
   Category(
     id: 'c1',
     title: 'Italian',
@@ -56,7 +56,7 @@ const DUMMY_CATEGORIES = const [
   ),
 ];
 
-const DUMMY_MEALS = const [
+const dummyMeals = [
   Meal(
     id: 'm1',
     categories: [
@@ -64,8 +64,8 @@ const DUMMY_MEALS = const [
       'c2',
     ],
     title: 'Spaghetti with Tomato Sauce',
-    affordability: Affordability.Affordable,
-    complexity: Complexity.Simple,
+    affordability: Affordability.affordable,
+    complexity: Complexity.simple,
     imageUrl:
         'https://upload.wikimedia.org/wikipedia/commons/thumb/2/20/Spaghetti_Bolognese_mit_Parmesan_oder_Grana_Padano.jpg/800px-Spaghetti_Bolognese_mit_Parmesan_oder_Grana_Padano.jpg',
     duration: 20,
@@ -97,8 +97,8 @@ const DUMMY_MEALS = const [
       'c2',
     ],
     title: 'Toast Hawaii',
-    affordability: Affordability.Affordable,
-    complexity: Complexity.Simple,
+    affordability: Affordability.affordable,
+    complexity: Complexity.simple,
     imageUrl:
         'https://cdn.pixabay.com/photo/2018/07/11/21/51/toast-3532016_1280.jpg',
     duration: 10,
@@ -126,8 +126,8 @@ const DUMMY_MEALS = const [
       'c3',
     ],
     title: 'Classic Hamburger',
-    affordability: Affordability.Pricey,
-    complexity: Complexity.Simple,
+    affordability: Affordability.pricey,
+    complexity: Complexity.simple,
     imageUrl:
         'https://cdn.pixabay.com/photo/2014/10/23/18/05/burger-500054_1280.jpg',
     duration: 45,
@@ -157,8 +157,8 @@ const DUMMY_MEALS = const [
       'c4',
     ],
     title: 'Wiener Schnitzel',
-    affordability: Affordability.Luxurious,
-    complexity: Complexity.Challenging,
+    affordability: Affordability.luxurious,
+    complexity: Complexity.challenging,
     imageUrl:
         'https://cdn.pixabay.com/photo/2018/03/31/19/29/schnitzel-3279045_1280.jpg',
     duration: 60,
@@ -194,8 +194,8 @@ const DUMMY_MEALS = const [
       'c10',
     ],
     title: 'Salad with Smoked Salmon',
-    affordability: Affordability.Luxurious,
-    complexity: Complexity.Simple,
+    affordability: Affordability.luxurious,
+    complexity: Complexity.simple,
     imageUrl:
         'https://cdn.pixabay.com/photo/2016/10/25/13/29/smoked-salmon-salad-1768890_1280.jpg',
     duration: 15,
@@ -229,8 +229,8 @@ const DUMMY_MEALS = const [
       'c10',
     ],
     title: 'Delicious Orange Mousse',
-    affordability: Affordability.Affordable,
-    complexity: Complexity.Hard,
+    affordability: Affordability.affordable,
+    complexity: Complexity.hard,
     imageUrl:
         'https://cdn.pixabay.com/photo/2017/05/01/05/18/pastry-2274750_1280.jpg',
     duration: 240,
@@ -264,8 +264,8 @@ const DUMMY_MEALS = const [
       'c7',
     ],
     title: 'Pancakes',
-    affordability: Affordability.Affordable,
-    complexity: Complexity.Simple,
+    affordability: Affordability.affordable,
+    complexity: Complexity.simple,
     imageUrl:
         'https://cdn.pixabay.com/photo/2018/07/10/21/23/pancake-3529653_1280.jpg',
     duration: 20,
@@ -295,8 +295,8 @@ const DUMMY_MEALS = const [
       'c8',
     ],
     title: 'Creamy Indian Chicken Curry',
-    affordability: Affordability.Pricey,
-    complexity: Complexity.Challenging,
+    affordability: Affordability.pricey,
+    complexity: Complexity.challenging,
     imageUrl:
         'https://cdn.pixabay.com/photo/2018/06/18/16/05/indian-food-3482749_1280.jpg',
     duration: 35,
@@ -328,8 +328,8 @@ const DUMMY_MEALS = const [
       'c9',
     ],
     title: 'Chocolate Souffle',
-    affordability: Affordability.Affordable,
-    complexity: Complexity.Hard,
+    affordability: Affordability.affordable,
+    complexity: Complexity.hard,
     imageUrl:
         'https://cdn.pixabay.com/photo/2014/08/07/21/07/souffle-412785_1280.jpg',
     duration: 45,
@@ -379,8 +379,8 @@ const DUMMY_MEALS = const [
       'c10',
     ],
     title: 'Asparagus Salad with Cherry Tomatoes',
-    affordability: Affordability.Luxurious,
-    complexity: Complexity.Simple,
+    affordability: Affordability.luxurious,
+    complexity: Complexity.simple,
     imageUrl:
         'https://cdn.pixabay.com/photo/2018/04/09/18/26/asparagus-3304997_1280.jpg',
     duration: 30,
