@@ -6,12 +6,12 @@ import '../widgets/meal_item.dart';
 class FavoritesScreen extends StatelessWidget {
   final List<Meal> favoriteMeals;
 
-  FavoritesScreen(this.favoriteMeals);
+  const FavoritesScreen(this.favoriteMeals, {super.key});
 
   @override
   Widget build(BuildContext context) {
     if (favoriteMeals.isEmpty) {
-      return Center(
+      return const Center(
         child: Text('You have no favorites yet - start adding some!'),
       );
     } else {
